PoC Wayland/libei support
This is the absolute minimum to check whether it works. The files are a copy
of the various XWindows*.(h|cpp) source files, with most of the stuff ifdefd
out or removed and replaced by FIXMEs.

This is only a client-side implementation and it's a compile-time option.
The server needs to run on XWindows to capture events.

It uses libei to communicate to an EIS implementation (read: the compositor)
and pass on the events. Think of libei replacing XTest, that's basically the
only real change here.

libei is a library for input emulation only, a number of things need to be
fetched from Wayland in this model.